refactor: 도메인을 패키지 별로 정리
