+++ src/main/java/kitchenpos/domain/product/Product.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.domain.product;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -6,17 +6,19 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 
 @Entity
-public class MenuGroup {
+public class Product {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private String name;
+    private ProductPrice productPrice;
 
-    public MenuGroup() {
+    public Product() {
     }
 
-    public MenuGroup(String name) {
+    public Product(String name, ProductPrice productPrice) {
         this.name = name;
+        this.productPrice = productPrice;
     }
 
     public Long getId() {
@@ -25,5 +27,9 @@ public class MenuGroup {
 
     public String getName() {
         return name;
+    }
+
+    public ProductPrice getProductPrice() {
+        return productPrice;
     }
 }
