+++ src/main/java/kitchenpos/domain/product/ProductPrice.java
@@ -1,6 +1,6 @@
-package kitchenpos.domain;
+package kitchenpos.domain.product;
 
-import kitchenpos.exception.InvalidMenuPriceException;
+import kitchenpos.exception.InvalidProductPriceException;
 
 import javax.persistence.Column;
 import javax.persistence.Embeddable;
@@ -8,26 +8,26 @@ import java.math.BigDecimal;
 import java.util.Objects;
 
 @Embeddable
-public class MenuPrice {
+public class ProductPrice {
     @Column(name = "price")
     private BigDecimal value;
 
-    public MenuPrice() {
+    public ProductPrice() {
     }
 
-    public MenuPrice(BigDecimal value) {
+    public ProductPrice(BigDecimal value) {
         validateValue(value);
         this.value = value;
     }
 
     private void validateValue(BigDecimal value) {
         if (Objects.isNull(value) || value.compareTo(BigDecimal.ZERO) < 0) {
-            throw new InvalidMenuPriceException("메뉴의 가격은 0원 이상이어야 합니다!");
+            throw new InvalidProductPriceException("상품의 가격은 0원 이상이어야 합니다!");
         }
     }
 
-    public boolean isBiggerThan(BigDecimal sum) {
-        return this.value.compareTo(sum) > 0;
+    public BigDecimal multiply(BigDecimal value) {
+        return this.value.multiply(value);
     }
 
     public BigDecimal getValue() {
