+++ src/main/java/kitchenpos/dto/menugroup/MenuGroupResponse.java
@@ -1,6 +1,6 @@
 package kitchenpos.dto.menugroup;
 
-import kitchenpos.domain.MenuGroup;
+import kitchenpos.domain.menu.MenuGroup;
 
 public class MenuGroupResponse {
     private Long id;
