+++ src/main/java/kitchenpos/dto/menuproduct/MenuProductCreateRequest.java
@@ -1,7 +1,7 @@
 package kitchenpos.dto.menuproduct;
 
 import kitchenpos.domain.MenuProduct;
-import kitchenpos.domain.Product;
+import kitchenpos.domain.product.Product;
 
 import javax.validation.constraints.NotNull;
