+++ src/main/java/kitchenpos/dto/product/ProductResponse.java
@@ -1,6 +1,6 @@
 package kitchenpos.dto.product;
 
-import kitchenpos.domain.Product;
+import kitchenpos.domain.product.Product;
 
 import java.math.BigDecimal;
