+++ src/main/java/kitchenpos/repository/OrderLineItemRepository.java
@@ -1,6 +1,6 @@
 package kitchenpos.repository;
 
-import kitchenpos.domain.OrderLineItem;
+import kitchenpos.domain.order.OrderLineItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
