+++ src/main/java/kitchenpos/repository/OrderRepository.java
@@ -1,7 +1,7 @@
 package kitchenpos.repository;
 
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderStatus;
+import kitchenpos.domain.order.Order;
+import kitchenpos.domain.order.OrderStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
