+++ src/main/java/kitchenpos/repository/TableGroupRepository.java
@@ -1,6 +1,6 @@
 package kitchenpos.repository;
 
-import kitchenpos.domain.TableGroup;
+import kitchenpos.domain.table.TableGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TableGroupRepository extends JpaRepository<TableGroup, Long> {
